add gitops check
